requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.28.0
scipy>=1.12.0
pandas>=2.2.0
matplotlib>=3.8.0
